@@ -1,99 +1,49 @@
 name: kinbo
-description: A new Flutter project.
-
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+description: "A new Flutter project."
+publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ^3.6.2
+
 
 dependencies:
-  cloud_firestore: ^2.4.0
-  cupertino_icons: ^1.0.3
-  firebase_auth: ^3.0.0
-  firebase_core: ^1.4.0
-  firebase_storage: ^10.0.0
   flutter:
     sdk: flutter
-  flutter_cache_manager: ^3.1.2
-  flutter_launcher_icons: ^0.9.1
-  flutter_spinkit: ^5.0.0
-  google_fonts: ^2.1.0
-  google_maps_flutter: ^2.0.6
-  image: ^3.0.2
-  image_picker: ^0.8.3
-  location: ^4.3.0
-  paginate_firestore: ^1.0.2
-  provider: ^5.0.0
+
+  cloud_firestore: ^5.6.3
+  cupertino_icons: ^1.0.6
+  firebase_auth: ^5.4.2
+  firebase_core: ^3.11.0
+  firebase_storage: ^12.4.2
+  flutter_cache_manager: ^3.3.1
+  flutter_launcher_icons: ^0.14.3
+  flutter_spinkit: ^5.2.0
+  google_fonts: ^6.2.1
+  google_maps_flutter: ^2.5.3
+  image: ^4.1.7
+  image_picker: ^1.0.7
+  location: ^8.0.0
+  #paginate_firestore: ^1.0.3
+  provider: ^6.1.2
+  intl: ^0.18.1
+  permission_handler: ^11.1.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-flutter_icons:
-  android: true
-  ios: true
-  image_path: "assets/icon.jpg"
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-# The following section is specific to Flutter.
+  flutter_lints: ^5.0.0
+
+
 flutter:
 
-  # The following line ensures that the Material Icons font is
-  # cloud_firestore: 
-  # cupertino_icons: ^1.0.2
-  # firebase_auth: ^0.20.1
-  # firebase_core: 0.7.0
-  # firebase_storage: 
-  # flutter_cache_manager: 
-  # flutter_spinkit: 
-  # google_maps_flutter:
-  # image: ^3.0.2
-  # image_picker:
-  # location:
-  # provider: ^4.3.2+1
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
+
   uses-material-design: true
-  # To add assets to your application, add an assets section, like this:
-  assets:
-    - markerBorder.png
-    - google_font/
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
+
+flutter_launcher_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "lib/assets/icon.jpg"
+  min_sdk_android: 23 # android min sdk min:16, default 21
