@@ -1,9 +1,9 @@
 import 'package:google_maps_flutter/google_maps_flutter.dart';
-import 'package:kinbo/model/user.dart';
+// import 'package:kinbo/model/user.dart';
 import 'package:location/location.dart';
 
 class UserLocationData {
-  LocationData myLocation;
+  LocationData ? myLocation;
 
   UserLocationData({this.myLocation});
 
