@@ -3,8 +3,8 @@ import 'package:kinbo/model/buddy.dart';
 import 'package:kinbo/model/user.dart';
 
 class DatabaseService {
-  final String uid;
-  String query;
+  final String ? uid;
+  String ? query;
 
   DatabaseService({this.uid, this.query});
 
@@ -14,13 +14,13 @@ class DatabaseService {
 
   //updates specific user data.
   Future updateUserData(
-      {String uuid,
-      String name,
-      String bio,
-      GeoPoint location,
-      String time,
-      List friends,
-      String image}) async {
+      {String ? uuid,
+      String ? name,
+      String ? bio,
+      GeoPoint ? location,
+      String ? time,
+      List ? friends,
+      String ? image}) async {
     return await buddyCollection.doc(uid).update({
       if (name != null) 'name': name,
       if (bio != null) 'bio': bio,
@@ -34,13 +34,13 @@ class DatabaseService {
 
   // Sets new user data during registration
   Future setNewUserData(
-      {String uuid,
-      String name,
-      String bio,
-      GeoPoint location,
-      String time,
-      List friends,
-      String image}) async {
+      {String ? uuid,
+      String ? name,
+      String ? bio,
+      GeoPoint ? location,
+      String ? time,
+      List ? friends,
+      String ? image}) async {
     return await buddyCollection.doc(uid).set({
       'name': name,
       'bio': bio,
@@ -55,19 +55,21 @@ class DatabaseService {
 //
 //--------------------------------------------------------------------
 // Create list of Buddy model locally from snapshot.
+
   List<Buddy> _buddyListFromSnapshot(QuerySnapshot snapshot) {
     return snapshot.docs.map((doc) {
       return Buddy(
-              uid: doc.id,
-              name: doc['name'] ?? null,
-              bio: doc['bio'] ?? null,
-              location: doc['location'] ?? null,
-              time: doc['time'] ?? null,
-              image: doc['image'] ?? null,
-              friends: doc['friends']) ??
-          null;
+        uid: doc.id,
+        name: doc['name'] as String?,
+        bio: doc['bio'] as String?,
+        location: doc['location'] as GeoPoint?,
+        time: doc['time'] as String?,
+        image: doc['image'] as String?,
+        friends: List.from(doc['friends'] ?? []),
+      );
     }).toList();
   }
+
 
   // Stream to notify for only searched user's update.
   Stream<List<Buddy>> get buddySearch {
@@ -86,56 +88,95 @@ class DatabaseService {
 //
 //-------------------------------------------------------------
 // Create UserData model locally
-  UserData userDataFromSnapshot(DocumentSnapshot doc) {
-    if (doc != null) {
+  UserData ? userDataFromSnapshot(DocumentSnapshot doc) {
+    if (doc.exists) {
       return UserData(
-              uid: uid,
-              name: doc['name'] ?? null,
-              bio: doc['bio'] ?? null,
-              location: doc['location'] ?? null,
-              image: doc['image'] ?? null,
-              time: doc['time'] ?? null,
-              friends: doc['friends']) ??
-          null;
+        uid: doc.id,
+        name: doc['name'] as String?,
+        bio: doc['bio'] as String?,
+        location: doc['location'] as GeoPoint?,
+        image: doc['image'] as String?,
+        time: doc['time'] as String?,
+        friends: List.from(doc['friends'] ?? []),
+      );
     } else {
       return null;
     }
   }
 
+
   // Stream for a specific user data.
+  // Stream <UserData?> get userData {
+  //   return buddyCollection.doc(uid).snapshots().map(userDataFromSnapshot);
+  // }
+
   Stream<UserData> get userData {
-    return buddyCollection.doc(uid).snapshots().map(userDataFromSnapshot);
+    return buddyCollection
+        .doc(uid)
+        .snapshots()
+        .map((doc) => userDataFromSnapshot(doc) ?? UserData());
   }
 
 //
 //--------------------------------------------------------------------
 // Create Friend model locally from snapshot.
+//   List<Friend?> _friendListFromSnapshot(QuerySnapshot snapshot) {
+//     return snapshot.docs.map((doc) {
+//       return Friend(
+//               uid: doc.id,
+//               name: doc['name'] ?? null,
+//               bio: doc['bio'] ?? null,
+//               location: doc['location'] ?? null,
+//               time: doc['time'] ?? null,
+//               image: doc['image'] ?? null,
+//               friends: doc['friends']) ??
+//           null;
+//     }).toList();
+//   }
+
   List<Friend> _friendListFromSnapshot(QuerySnapshot snapshot) {
     return snapshot.docs.map((doc) {
+      if (doc.data() == null) return null;
+
       return Friend(
-              uid: doc.id,
-              name: doc['name'] ?? null,
-              bio: doc['bio'] ?? null,
-              location: doc['location'] ?? null,
-              time: doc['time'] ?? null,
-              image: doc['image'] ?? null,
-              friends: doc['friends']) ??
-          null;
-    }).toList();
+        uid: doc.id,
+        name: doc['name'] ?? '',
+        bio: doc['bio'] ?? '',
+        location: doc['location'] ?? '',
+        time: doc['time'] ?? '',
+        image: doc['image'] ?? '',
+        friends: (doc['friends'] as List<dynamic>?)?.cast<String>() ?? [],
+      );
+    }).whereType<Friend>().toList();
   }
 
+
+
   // Stream to notify for only friends' update.
+  // Stream<List<Friend?>> get friends {
+  //   return buddyCollection
+  //       .where('friends', arrayContainsAny: [this.uid])
+  //       .snapshots()
+  //       .map(_friendListFromSnapshot);
+  // }
+
   Stream<List<Friend>> get friends {
     return buddyCollection
         .where('friends', arrayContainsAny: [this.uid])
         .snapshots()
-        .map(_friendListFromSnapshot);
+        .map((snapshot) {
+      return _friendListFromSnapshot(snapshot)
+          .whereType<Friend>()
+          .toList();
+    });
   }
+
+
 
 //
 //----------------------------------------------------------------------
 // Future method to initiate friends' marker icons.
-  Future<List<Friend>> buddyImages() {
+  Future<List<Friend?>> buddyImages() {
     return buddyCollection
         .where('friends', arrayContainsAny: [this.uid])
         .get()
