@@ -1,18 +1,13 @@
-import UIKit
 import Flutter
-import GoogleMaps
+import UIKit
 
-@UIApplicationMain
+@main
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
     didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
   ) -> Bool {
     GeneratedPluginRegistrant.register(with: self)
-    
-    GMSServices.provideAPIKey("AIzaSyD5K7UMi0YUQrecYUc-uRokYFiKbOugVe0")
-
     return super.application(application, didFinishLaunchingWithOptions: launchOptions)
   }
 }
-
