@@ -1,3 +1,4 @@
+
 import 'package:flutter/material.dart';
 import 'package:kinbo/model/buddy.dart';
 import 'package:kinbo/model/user.dart';
@@ -8,15 +9,9 @@ import 'package:kinbo/views/user_panel/people/find_friends.dart';
 import 'package:kinbo/views/user_panel/profile/profile.dart';
 import 'package:provider/provider.dart';
 
-
-// two tabs to switch between profile and find friend. 
+// two tabs to switch between profile and find friend.
 
 class UserPanel extends StatefulWidget {
-  // final String uid;
-
-  //UserPanel(this.uid);
-  //const UserPanel({ Key? key }) : super(key: key);
-
   @override
   _UserPanelState createState() => _UserPanelState();
 }
@@ -24,28 +19,24 @@ class UserPanel extends StatefulWidget {
 class _UserPanelState extends State<UserPanel> {
   @override
   Widget build(BuildContext context) {
-    return MultiProvider(
-      providers: [
-        // Providers without UID needed.
-        StreamProvider<UserObj>.value(
-            initialData: UserObj(), value: AuthService().user),
-        // StreamProvider<List<Buddy>>.value(
-        //     initialData: [], value: DatabaseService().buddies),
-        //StreamProvider<LocationData>.value(value: LocationService().getLocation),
-      ],
-      child: Builder(
-        builder: (context) {
-          final user = Provider.of<UserObj>(context);
-          if (user != null) {
+    return StreamProvider<UserObj?>.value(
+      initialData: null,
+      value: AuthService().user,
+      child: Consumer<UserObj?>(
+        builder: (context, user, child) {
+          if (user == null) {
+            return Loading();
+          } else {
             return MultiProvider(
               providers: [
                 // Providers with UID needed.
                 StreamProvider<List<Friend>>.value(
-                    initialData: null,
+                    initialData: [],
                     value: DatabaseService(uid: user.uid).friends),
-                StreamProvider<UserData>.value(
-                    initialData: null,
-                    value: DatabaseService(uid: user.uid).userData),
+                StreamProvider<UserData?>.value(
+                  initialData: null,
+                  value: DatabaseService(uid: user.uid).userData,
+                ),
               ],
               child: DefaultTabController(
                 length: 2,
@@ -66,22 +57,24 @@ class _UserPanelState extends State<UserPanel> {
                     ),
                     backgroundColor: Colors.indigo[300],
                   ),
-                  // title: const Text('Tabs Demo'),
-
                   body: TabBarView(
                     children: [
-                      Profile(user.uid),
+                      Consumer<UserData?>(
+                        builder: (context, userData, child) {
+                          return Profile(user.uid??'');
+                        },
+                      ),
                       FindFriends(),
                     ],
                   ),
                 ),
               ),
             );
-          } else {
-            return Loading();
           }
         },
       ),
     );
   }
 }
+
+
